Moved some data into the main class; added a couple more tests + actions
The engine now grabs vocab, controls and messages from the data object
before it gets passed to the game state. This means it no longer has to
reach through the state object to reference data in the data object.

Also some other restructuring; for example, words is now a class member
so it doesn't have to get passed all through the stack every turn.