Change word matching strategy
Added a dict of synonyms, where every word has a list of words that are
its synonyms. This replaces the old vocab which was a bunch of unindexed
lists. This is better for a few reasons, efficiency being one of them.

Tightened up the vocab function a bit, but I'll probably remove it as
it's no longer used.