PLAYABLE. Fixed bugs, refactored data functions, altered game logic
The game now seems to be playable to the end; finished the game without
running into any showstoppers last time around.

The noun location functions now take an arbitrary number of arguments.

Also, made the first change to Starflight's game logic, as it appeared
to be broken in the original PHP/XML version as well (see note in
previous commit).