Added import/export state, aka "save game"
Allows for extraction of the minimal amount of data necessary to
reconstruct the game at a certain state.

Note: this doesn't yet work in Flask because the state variable is still
too large to fit in a client-side cookie, which carries a max of 4K of
data. Next step is to look into server-side storage.