Added wildcard filters, fixing bugs in examine
As of the last commit, the examine command in Starflight checked if any
of the matching nouns were present, then printed all matching nouns'
descriptions, which was wrong. To fix this, a filter feature has been
added.

By appending :test to the end of a wildcard, where 'test' is the name
of a t_ method in the tests class that has the @is_filter decorator, the
matching nouns will be filtered based on whether they pass that test.